+++ sources/api/netdog/test_data/net_config/basic/net_config.toml
@@ -1,0 +1,50 @@
+version = {{version}}
+
+[eno1]
+dhcp4 = true
+
+[eno2]
+dhcp6 = true
+primary = true
+
+[eno3]
+dhcp4 = true
+dhcp6 = false
+
+[eno4]
+dhcp4 = false
+dhcp6 = true
+
+[eno5]
+dhcp4 = true
+dhcp6 = true
+
+[eno6.dhcp4]
+enabled = true
+route-metric = 100
+
+[eno6]
+dhcp6 = false
+
+[eno7]
+dhcp4 = true
+
+[eno7.dhcp6]
+enabled = true
+optional = true
+
+[eno8.dhcp4]
+enabled = true
+optional = true
+
+[eno8.dhcp6]
+enabled = true
+optional = true
+
+[eno9.dhcp4]
+enabled = true
+optional = true
+
+[eno10.dhcp6]
+enabled = true
+optional = true
