+++ sources/api/netdog/test_data/net_config/basic/no_interfaces.toml
@@ -1,0 +1,1 @@
+version = {{version}}
