+++ sources/api/netdog/test_data/net_config/basic/no_primary.toml
@@ -1,4 +1,4 @@
-version = 1
+version = {{version}}
 
 [eno3]
 dhcp4 = true
