+++ sources/api/netdog/test_data/net_config/dhcp/invalid_dhcp6_options.toml
@@ -1,4 +1,4 @@
-version = 1
+version = {{version}}
 
 [eno1]
 dhcp4 = true
