+++ sources/api/netdog/test_data/net_config/dhcp/invalid_dhcp_config.toml
@@ -1,0 +1,3 @@
+version = {{version}}
+
+[eno1]
