--- sources/api/netdog/test_data/net_config/invalid_dhcp_config.toml
@@ -1,3 +1,0 @@
-version = 1
-
-[eno1]
